Move tests under the ORM namespace.
