+++ tests/TestCase/Database/QueryTests/CaseExpressionQueryTest.php
@@ -301,95 +301,4 @@ class CaseExpressionQueryTest extends TestCase
         ];
         $this->assertSame($expected, $query->execute()->fetchAll(StatementInterface::FETCH_TYPE_ASSOC));
     }
-
-    public function testOverwrittenReturnType(): void
-    {
-        $typeMap = new TypeMap([
-            'is_cheap' => 'boolean',
-        ]);
-
-        $query = $this->query
-            ->select(function (Query $query) {
-                return [
-                    'name',
-                    'price',
-                    'is_cheap' => $query->newExpr()
-                        ->case()
-                        ->when(['price <' => 20])
-                        ->then(1)
-                        ->else(0)
-                        ->setReturnType('boolean'),
-                ];
-            })
-            ->from('products')
-            ->orderAsc('price')
-            ->orderAsc('name')
-            ->setSelectTypeMap($typeMap);
-
-        $expected = [
-            [
-                'name' => 'First product',
-                'price' => '10',
-                'is_cheap' => true,
-            ],
-            [
-                'name' => 'Second product',
-                'price' => '20',
-                'is_cheap' => false,
-            ],
-            [
-                'name' => 'Third product',
-                'price' => '30',
-                'is_cheap' => false,
-            ],
-        ];
-        $this->assertSame($expected, $query->execute()->fetchAll(StatementInterface::FETCH_TYPE_ASSOC));
-    }
-
-    public function bindingValueDataProvider(): array
-    {
-        return [
-            ['1', 3],
-            ['2', 4],
-        ];
-    }
-
-    /**
-     * @dataProvider bindingValueDataProvider
-     * @param string $when The `WHEN` value.
-     * @param int $result The result value.
-     */
-    public function testBindValues(string $when, int $result): void
-    {
-        $value = '1';
-        $then = '3';
-        $else = '4';
-
-        $typeMap = new TypeMap([
-            'val' => 'integer',
-        ]);
-
-        $query = $this->query
-            ->select(function (Query $query) {
-                return [
-                    'val' => $query->newExpr()
-                        ->case($query->newExpr(':value'))
-                        ->when($query->newExpr(':when'))
-                        ->then($query->newExpr(':then'))
-                        ->else($query->newExpr(':else'))
-                        ->setReturnType('integer'),
-                ];
-            })
-            ->from('products')
-            ->bind(':value', $value, 'integer')
-            ->bind(':when', $when, 'integer')
-            ->bind(':then', $then, 'integer')
-            ->bind(':else', $else, 'integer')
-            ->setSelectTypeMap($typeMap);
-
-        $expected = [
-            'val' => $result,
-        ];
-        $this->assertSame($expected, $query->execute()->fetch(StatementInterface::FETCH_TYPE_ASSOC));
-    }
 }
